refactor: split register constants
