+++ custom_components/thessla_green_modbus/registers.py
@@ -1,0 +1,282 @@
+"""Register definitions for the ThesslaGreen Modbus integration."""
+from typing import Dict
+
+# INPUT REGISTERS (04 - READ INPUT REGISTER) - Sensors and read-only values
+INPUT_REGISTERS: Dict[str, int] = {
+    # Firmware version (0x0000-0x0004)
+    "firmware_major": 0x0000,
+    "firmware_minor": 0x0001,
+    "day_of_week": 0x0002,
+    "period": 0x0003,
+    "firmware_patch": 0x0004,
+    # Compilation info (0x000E-0x000F)
+    "compilation_days": 0x000E,
+    "compilation_seconds": 0x000F,
+    # Temperature sensors (0x0010-0x0017) - 0.1°C resolution, 0x8000 = no sensor
+    "outside_temperature": 0x0010,
+    "supply_temperature": 0x0011,
+    "exhaust_temperature": 0x0012,
+    "fpx_temperature": 0x0013,
+    "duct_supply_temperature": 0x0014,
+    "gwc_temperature": 0x0015,
+    "ambient_temperature": 0x0016,
+    "heating_temperature": 0x0017,
+    # Flow sensors (0x0018-0x001E) - m³/h, 0x8000 = no sensor
+    "supply_flowrate": 0x0018,
+    "exhaust_flowrate": 0x0019,
+    "outdoor_flowrate": 0x001A,
+    "inside_flowrate": 0x001B,
+    "gwc_flowrate": 0x001C,
+    "heat_recovery_flowrate": 0x001D,
+    "bypass_flowrate": 0x001E,
+    # Air quality sensors (0x0020-0x0027)
+    "co2_level": 0x0020,
+    "humidity_indoor": 0x0021,
+    "humidity_outdoor": 0x0022,
+    "pm1_level": 0x0023,
+    "pm25_level": 0x0024,
+    "pm10_level": 0x0025,
+    "voc_level": 0x0026,
+    "air_quality_index": 0x0027,
+    # System status registers (0x0030-0x003F)
+    "heat_recovery_efficiency": 0x0030,
+    "filter_lifetime_remaining": 0x0031,
+    "preheater_power": 0x0032,
+    "main_heater_power": 0x0033,
+    "cooler_power": 0x0034,
+    "supply_fan_power": 0x0035,
+    "exhaust_fan_power": 0x0036,
+    "total_power_consumption": 0x0037,
+    "annual_energy_consumption": 0x0038,
+    "daily_energy_consumption": 0x0039,
+    "annual_energy_savings": 0x003A,
+    "co2_reduction": 0x003B,
+    "system_uptime": 0x003C,
+    "fault_counter": 0x003D,
+    "maintenance_counter": 0x003E,
+    "filter_replacement_counter": 0x003F,
+    # Expansion module version (0x00F1)
+    "expansion_version": 0x00F1,
+    # Current airflow (0x0100-0x0101)
+    "supply_air_flow": 0x0100,
+    "exhaust_air_flow": 0x0101,
+    # PWM control values (0x0500-0x0503) - 0-4095 (0-10V)
+    "dac_supply": 0x0500,
+    "dac_exhaust": 0x0501,
+    "dac_heater": 0x0502,
+    "dac_cooler": 0x0503,
+    # Advanced diagnostics (0x0504-0x051F)
+    "motor_supply_rpm": 0x0504,
+    "motor_exhaust_rpm": 0x0505,
+    "motor_supply_current": 0x0506,
+    "motor_exhaust_current": 0x0507,
+    "motor_supply_voltage": 0x0508,
+    "motor_exhaust_voltage": 0x0509,
+    "supply_pressure": 0x050A,
+    "exhaust_pressure": 0x050B,
+    "differential_pressure": 0x050C,
+    "heat_exchanger_temperature_1": 0x050D,
+    "heat_exchanger_temperature_2": 0x050E,
+    "heat_exchanger_temperature_3": 0x050F,
+    "heat_exchanger_temperature_4": 0x0510,
+    "damper_position_bypass": 0x0511,
+    "damper_position_gwc": 0x0512,
+    "damper_position_mix": 0x0513,
+    "frost_protection_active": 0x0514,
+    "defrost_cycle_active": 0x0515,
+    "summer_bypass_active": 0x0516,
+    "winter_heating_active": 0x0517,
+    "night_cooling_active": 0x0518,
+    "constant_flow_active": 0x0519,
+    "air_quality_control_active": 0x051A,
+    "humidity_control_active": 0x051B,
+    "temperature_control_active": 0x051C,
+    "demand_control_active": 0x051D,
+    "schedule_control_active": 0x051E,
+    "manual_control_active": 0x051F,
+}
+
+# HOLDING REGISTERS (03 - READ/WRITE HOLDING REGISTER) - Configuration and control
+HOLDING_REGISTERS: Dict[str, int] = {
+    # Main control registers (0x1000-0x1020)
+    "on_off_panel_mode": 0x1000,
+    "mode": 0x1001,
+    "air_flow_rate_manual": 0x1002,
+    "supply_air_temperature_manual": 0x1003,
+    "special_mode": 0x1004,
+    "required_temperature": 0x1005,
+    "comfort_temperature": 0x1006,
+    "economy_temperature": 0x1007,
+    "night_temperature": 0x1008,
+    "away_temperature": 0x1009,
+    "frost_protection_temperature": 0x100A,
+    "max_supply_temperature": 0x100B,
+    "min_supply_temperature": 0x100C,
+    "heating_curve_slope": 0x100D,
+    "heating_curve_offset": 0x100E,
+    "supply_air_flow_rate": 0x100F,
+    "exhaust_air_flow_rate": 0x1010,
+    "minimum_air_flow_rate": 0x1011,
+    "maximum_air_flow_rate": 0x1012,
+    "eco_air_flow_rate": 0x1013,
+    "night_air_flow_rate": 0x1014,
+    "away_air_flow_rate": 0x1015,
+    "boost_air_flow_rate": 0x1016,
+    "boost_duration": 0x1017,
+    "fireplace_air_flow_rate": 0x1018,
+    "fireplace_duration": 0x1019,
+    "hood_air_flow_rate": 0x101A,
+    "hood_duration": 0x101B,
+    "party_air_flow_rate": 0x101C,
+    "party_duration": 0x101D,
+    "bathroom_air_flow_rate": 0x101E,
+    "bathroom_duration": 0x101F,
+    "kitchen_air_flow_rate": 0x1020,
+    # Special functions (0x1030-0x1050)
+    "bypass_mode": 0x1030,
+    "bypass_temperature_threshold": 0x1031,
+    "bypass_hysteresis": 0x1032,
+    "gwc_mode": 0x1033,
+    "gwc_temperature_threshold": 0x1034,
+    "gwc_hysteresis": 0x1035,
+    "preheating_mode": 0x1036,
+    "preheating_temperature": 0x1037,
+    "defrost_mode": 0x1038,
+    "defrost_temperature": 0x1039,
+    "defrost_duration": 0x103A,
+    "frost_protection_mode": 0x103B,
+    "summer_mode": 0x103C,
+    "winter_mode": 0x103D,
+    "night_cooling_mode": 0x103E,
+    "night_cooling_temperature": 0x103F,
+    "air_quality_control": 0x1040,
+    "humidity_control": 0x1041,
+    "humidity_target": 0x1042,
+    "humidity_hysteresis": 0x1043,
+    "demand_control": 0x1044,
+    "occupancy_sensor": 0x1045,
+    "window_contact": 0x1046,
+    "external_switch": 0x1047,
+    "bathroom_switch": 0x1048,
+    "kitchen_switch": 0x1049,
+    "bedroom_switch": 0x104A,
+    "living_room_switch": 0x104B,
+    "office_switch": 0x104C,
+    "guest_room_switch": 0x104D,
+    "basement_switch": 0x104E,
+    "attic_switch": 0x104F,
+    "garage_switch": 0x1050,
+    # Air quality settings (0x1060-0x1080)
+    "co2_threshold_low": 0x1060,
+    "co2_threshold_medium": 0x1061,
+    "co2_threshold_high": 0x1062,
+    "co2_hysteresis": 0x1063,
+    "voc_threshold_low": 0x1064,
+    "voc_threshold_medium": 0x1065,
+    "voc_threshold_high": 0x1066,
+    "voc_hysteresis": 0x1067,
+    "pm25_threshold_low": 0x1068,
+    "pm25_threshold_medium": 0x1069,
+    "pm25_threshold_high": 0x106A,
+    "pm25_hysteresis": 0x106B,
+    "pm10_threshold_low": 0x106C,
+    "pm10_threshold_medium": 0x106D,
+    "pm10_threshold_high": 0x106E,
+    "pm10_hysteresis": 0x106F,
+    "air_quality_response_delay": 0x1070,
+    "air_quality_boost_duration": 0x1071,
+    "air_quality_boost_flow": 0x1072,
+    "filter_alarm_threshold": 0x1073,
+    "filter_warning_threshold": 0x1074,
+    "maintenance_interval": 0x1075,
+    "cleaning_reminder": 0x1076,
+    "sensor_calibration_interval": 0x1077,
+    "backup_settings": 0x1078,
+    "factory_reset": 0x1079,
+    "configuration_lock": 0x107A,
+    "user_level": 0x107B,
+    "display_brightness": 0x107C,
+    "display_timeout": 0x107D,
+    "keypad_lock": 0x107E,
+    "sound_volume": 0x107F,
+    "language": 0x1080,
+    # Weekly schedule (0x1100-0x1127)
+    "weekly_schedule_mode": 0x1100,
+    "schedule_monday_period1_start": 0x1101,
+    "schedule_monday_period1_end": 0x1102,
+    "schedule_monday_period1_flow": 0x1103,
+    "schedule_monday_period1_temp": 0x1104,
+    "schedule_monday_period2_start": 0x1105,
+    "schedule_monday_period2_end": 0x1106,
+    "schedule_monday_period2_flow": 0x1107,
+    "schedule_monday_period2_temp": 0x1108,
+    "schedule_tuesday_period1_start": 0x1109,
+    "schedule_tuesday_period1_end": 0x110A,
+    "schedule_tuesday_period1_flow": 0x110B,
+    "schedule_tuesday_period1_temp": 0x110C,
+    "schedule_tuesday_period2_start": 0x110D,
+    "schedule_tuesday_period2_end": 0x110E,
+    "schedule_tuesday_period2_flow": 0x110F,
+    "schedule_tuesday_period2_temp": 0x1110,
+    "schedule_wednesday_period1_start": 0x1111,
+    "schedule_wednesday_period1_end": 0x1112,
+    "schedule_wednesday_period1_flow": 0x1113,
+    "schedule_wednesday_period1_temp": 0x1114,
+    "schedule_wednesday_period2_start": 0x1115,
+    "schedule_wednesday_period2_end": 0x1116,
+    "schedule_wednesday_period2_flow": 0x1117,
+    "schedule_wednesday_period2_temp": 0x1118,
+    "schedule_thursday_period1_start": 0x1119,
+    "schedule_thursday_period1_end": 0x111A,
+    "schedule_thursday_period1_flow": 0x111B,
+    "schedule_thursday_period1_temp": 0x111C,
+    "schedule_thursday_period2_start": 0x111D,
+    "schedule_thursday_period2_end": 0x111E,
+    "schedule_thursday_period2_flow": 0x111F,
+    "schedule_thursday_period2_temp": 0x1120,
+    "schedule_friday_period1_start": 0x1121,
+    "schedule_friday_period1_end": 0x1122,
+    "schedule_friday_period1_flow": 0x1123,
+    "schedule_friday_period1_temp": 0x1124,
+    "schedule_friday_period2_start": 0x1125,
+    "schedule_friday_period2_end": 0x1126,
+    "schedule_friday_period2_flow": 0x1127,
+    # Temporary mode control (0x1130-0x1135) - from PDF documentation
+    "cfg_mode1": 0x1130,
+    "air_flow_rate_temporary": 0x1131,
+    "airflow_rate_change_flag": 0x1132,
+    "cfg_mode2": 0x1133,
+    "supply_air_temperature_temporary": 0x1134,
+    "temperature_change_flag": 0x1135,
+    # System reset controls (0x113D-0x113E) - from PDF documentation
+    "hard_reset_settings": 0x113D,
+    "hard_reset_schedule": 0x113E,
+    # Filter control (0x1150-0x1151) - from PDF documentation
+    "pres_check_day": 0x1150,
+    "pres_check_time": 0x1151,
+    # Modbus communication settings (0x1164-0x116B) - from PDF documentation
+    "uart0_id": 0x1164,
+    "uart0_baud": 0x1165,
+    "uart0_parity": 0x1166,
+    "uart0_stop": 0x1167,
+    "uart1_id": 0x1168,
+    "uart1_baud": 0x1169,
+    "uart1_parity": 0x116A,
+    "uart1_stop": 0x116B,
+    # Device name (0x1FD0-0x1FD7) - from PDF documentation
+    "device_name_1": 0x1FD0,
+    "device_name_2": 0x1FD1,
+    "device_name_3": 0x1FD2,
+    "device_name_4": 0x1FD3,
+    "device_name_5": 0x1FD4,
+    "device_name_6": 0x1FD5,
+    "device_name_7": 0x1FD6,
+    "device_name_8": 0x1FD7,
+    # Product key and lock (0x1FFB-0x1FFF) - from PDF documentation
+    "lock_pass1": 0x1FFB,
+    "lock_pass2": 0x1FFC,
+    "lock_flag": 0x1FFD,
+    # Legacy location for required temperature used by older firmware
+    "required_temperature_legacy": 0x1FFE,
+    "filter_change": 0x1FFF,
+}
