+++ custom_components/thessla_green_modbus/switch.py
@@ -24,7 +24,8 @@ except (ModuleNotFoundError, ImportError):  # pragma: no cover
 
         pass
 
-from .const import COIL_REGISTERS, DOMAIN, HOLDING_REGISTERS
+from .const import COIL_REGISTERS, DOMAIN
+from .registers import HOLDING_REGISTERS
 from .coordinator import ThesslaGreenModbusCoordinator
 from .entity import ThesslaGreenEntity
