+++ tests/test_climate.py
@@ -101,11 +101,9 @@ sys.modules["homeassistant.helpers.device_registry"] = device_registry
 # ---------------------------------------------------------------------------
 
 from custom_components.thessla_green_modbus.climate import ThesslaGreenClimate
-from custom_components.thessla_green_modbus.const import (
-    DOMAIN,
-    HOLDING_REGISTERS,
-    REGISTER_MULTIPLIERS,
-)
+from custom_components.thessla_green_modbus.const import DOMAIN
+from custom_components.thessla_green_modbus.multipliers import REGISTER_MULTIPLIERS
+from custom_components.thessla_green_modbus.registers import HOLDING_REGISTERS
 from custom_components.thessla_green_modbus.coordinator import (
     ThesslaGreenModbusCoordinator,
 )
