+++ tests/test_services_scaling.py
@@ -5,7 +5,7 @@ from types import SimpleNamespace
 
 import pytest
 
-from custom_components.thessla_green_modbus.const import HOLDING_REGISTERS
+from custom_components.thessla_green_modbus.registers import HOLDING_REGISTERS
 import custom_components.thessla_green_modbus.services as services
 from custom_components.thessla_green_modbus.services import _scale_for_register
